# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/delegations:
    get:
      tags:
        - Delegations
      summary: List Delegations
      description: Get a paginated list of Delegations with optional filtering and sorting
      operationId: listDelegations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelegationList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Delegations
      summary: Create Delegation
      description: Create a new Delegation
      operationId: createDelegation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDelegationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Delegation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Delegation:
      type: object
      required:
        - id
        - company_id
        - approver_id
        - delegate_to_id
        - valid_from
        - valid_to
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        approver_id:
          type: string
          format: uuid
        delegate_to_id:
          type: string
          format: uuid
        valid_from:
          type: string
          format: date
        valid_to:
          type: string
          format: date
        reason:
          nullable: true
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreateDelegationRequest:
      type: object
      required:
        - company_id
        - approver_id
        - delegate_to_id
        - valid_from
        - valid_to
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        approver_id:
          type: string
          format: uuid
        delegate_to_id:
          type: string
          format: uuid
        valid_from:
          type: string
          format: date
        valid_to:
          type: string
          format: date
        reason:
          nullable: true
          type: string
        status:
          type: string
        metadata:
          type: object
...
    DelegationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Delegation'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []